Move PDFGenerateError to PDFGenerateError.swift and edit documentation
